+++ spec/spec_helper.rb
@@ -5,5 +5,5 @@ support_files = File.expand_path("spec/support/**/*.rb")
 Dir[support_files].each { |file| require file }
 
 RSpec.configure do |config|
-  config.include Fixture
+  config.include SpecSupport::Methods
 end