+++ spec/support/fixture.rb
@@ -1,20 +1,42 @@
-module Fixture
+module SpecSupport
 
-  def self.[](name)
-    fixtures[name]
+  class Fixture
+    def self.[](name)
+      fixtures[name]
+    end
+
+    def self.[]=(name, value)
+      fixtures[name] = value
+    end
+
+    def self.fixtures
+      @fixtures ||= {}
+    end
+
+    def initialize(file, ext = :wsdl)
+      self.file = file
+      self.ext = ext
+    end
+
+    attr_accessor :file, :ext
+
+    def filename
+      "#{file}.#{ext}"
+    end
+
+    def path
+      File.expand_path("spec/fixtures/#{filename}")
+    end
+
+    def read
+      Fixture[filename] ||= File.read(path)
+    end
   end
 
-  def self.[]=(name, value)
-    fixtures[name] = value
-  end
-
-  def self.fixtures
-    @fixtures ||= {}
-  end
-
-  def fixture(file, ext = :wsdl)
-    filename = "#{file}.#{ext}"
-    Fixture[filename] ||= File.read File.expand_path("spec/fixtures/#{filename}")
+  module Methods
+    def fixture(*args)
+      Fixture.new(*args)
+    end
   end
 
 end