+++ spec/wasabi/document_spec.rb
@@ -2,7 +2,26 @@ require "spec_helper"
 
 describe Wasabi::Document do
 
-  subject { Wasabi::Document.new }
+  subject { Wasabi::Document.new fixture(:authentication).read }
+
+  describe ".new" do
+    it "accepts a URL" do
+      HTTPI.expects(:get).returns(HTTPI::Response.new(200, {}, "wsdl"))
+
+      document = Wasabi::Document.new("http://example.com?wsdl")
+      document.xml.should == "wsdl"
+    end
+
+    it "accepts a path" do
+      document = Wasabi::Document.new fixture(:authentication).path
+      document.xml.should == fixture(:authentication).read
+    end
+
+    it "accepts raw XML" do
+      document = Wasabi::Document.new fixture(:authentication).read
+      document.xml.should == fixture(:authentication).read
+    end
+  end
 
   describe ".validate_element_form_default!" do
     [:unqualified, :qualified].each do |value|