+++ spec/wasabi/wasabi_spec.rb
@@ -4,7 +4,7 @@ describe Wasabi do
 
   describe ".document" do
     it "should return a new Wasabi::Document" do
-      document = Wasabi.document fixture(:authentication)
+      document = Wasabi.document fixture(:authentication).read
       document.should be_a(Wasabi::Document)
     end
   end